requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
tqdm
pyyamlpip